Refactor config module
The commit separates the config module into multiple sub-modules, enhancing clarity and maintainability. It splits the former single file into 'app_settings.rs' and 'scheduled_message.rs'.
