+++ r2d2wm-bot/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "r2d2wm-bot"
 authors = ["Noé Ksiazek <noe@ksiazek.xyz>"]
-version = "0.1.0"
+version = "0.2.0"
 edition = "2021"
 rust-version = "1.79"
