+++ r2d2wm-bot/src/config.rs
@@ -1,73 +1,15 @@
-use crate::Result;
-use serde::Deserialize;
-use serenity::all::{ChannelId, Context, CreateMessage, MessageBuilder, RoleId, UserId};
 use std::env;
-use std::num::NonZeroU64;
 use std::path::{Path, PathBuf};
-use std::sync::Arc;
+
+pub use app_settings::AppSettings;
+pub use scheduled_message::ScheduledMessage;
+
+use crate::Result;
+
+mod app_settings;
+mod scheduled_message;
 
 const ENV_CONFIG_PATH: &str = "R2D2WM_CONFIG_PATH";
-
-#[derive(Deserialize, Debug, PartialEq)]
-pub struct AppSettings {
-    pub discord_token: String,
-    pub logging_level: String,
-    pub timezone: String,
-}
-
-#[derive(Deserialize, Debug, PartialEq, Clone)]
-pub struct ScheduledMessage {
-    pub name: String,
-    pub cron: String,
-    pub channel_id: NonZeroU64,
-    pub mentions: Option<Vec<MentionTarget>>,
-    pub message: String,
-}
-
-impl ScheduledMessage {
-    pub async fn send(&self, ctx: Arc<Context>) {
-        let channel: ChannelId = ChannelId::new(self.channel_id.get());
-        let message: CreateMessage = self.build_discord_message();
-
-        match channel.send_message(&ctx.http, message.clone()).await {
-            Ok(msg) => {
-                tracing::info!("Message sent on schedule: {:?}", &msg.id);
-                tracing::trace!("{:?}", &msg);
-            }
-            Err(e) => {
-                tracing::error!("Failed to send message: {:?}: {:?}", &self.name, e);
-            }
-        }
-    }
-
-    fn build_discord_message(&self) -> CreateMessage {
-        let mut msg_builder: MessageBuilder = MessageBuilder::new();
-        msg_builder.push(&self.message);
-
-        if let Some(mentions) = &self.mentions {
-            msg_builder.push_line("");
-
-            for (i, target) in mentions.iter().enumerate() {
-                match target {
-                    MentionTarget::Role(id) => msg_builder.mention(&RoleId::new(id.get())),
-                    MentionTarget::User(id) => msg_builder.mention(&UserId::new(id.get())),
-                };
-                if i < mentions.len() - 1 {
-                    msg_builder.push(" ");
-                }
-            }
-        }
-
-        CreateMessage::new().content(msg_builder.build())
-    }
-}
-
-#[derive(Deserialize, Debug, PartialEq, Clone)]
-#[serde(tag = "type", content = "id", rename_all = "lowercase")]
-pub enum MentionTarget {
-    Role(NonZeroU64),
-    User(NonZeroU64),
-}
 
 #[derive(Debug)]
 pub struct Config {
@@ -90,7 +32,7 @@ impl Config {
         Ok(config)
     }
 
-    pub fn get_schedules_from_file() -> Result<Vec<ScheduledMessage>> {
+    fn get_schedules_from_file() -> Result<Vec<ScheduledMessage>> {
         let path = Self::construct_path_to("schedule.json");
         let data: String = std::fs::read_to_string(path)?;
         let schedules: Vec<ScheduledMessage> = serde_json::from_str(&data)?;
@@ -106,8 +48,11 @@ impl Config {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use std::num::NonZeroU64;
+
     use serde_json::json;
+
+    use super::*;
 
     fn app_settings() -> AppSettings {
         AppSettings {
