+++ r2d2wm-bot/src/config/app_settings.rs
@@ -1,0 +1,8 @@
+use serde::Deserialize;
+
+#[derive(Deserialize, Debug, PartialEq)]
+pub struct AppSettings {
+    pub discord_token: String,
+    pub logging_level: String,
+    pub timezone: String,
+}
